fixed the downloadCSV bug in the rebate admin
- created a new service filterService to provide filter
  function commands throughout the project.
